redirection_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
pydantic[email]==2.7.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
alembic==1.13.1